:truck: Rename grorg/ to config/
- Rename project directory from grorg to config
- Update all imports and references
- Add CLAUDE.md documentation
