+++ pyproject.toml
@@ -27,7 +27,7 @@ tag = true
 version_pattern = "YYYY.MM.INC1"
 
 [tool.bumpver.file_patterns]
-"grorg/__init__.py" = [
+"config/__init__.py" = [
     '__version__ = "{version}"'
 ]
 "pyproject.toml" = [
@@ -35,7 +35,7 @@ version_pattern = "YYYY.MM.INC1"
 ]
 
 [tool.pytest.ini_options]
-DJANGO_SETTINGS_MODULE = "grorg.settings"
+DJANGO_SETTINGS_MODULE = "config.settings"
 addopts = "--nomigrations --reuse-db"
 norecursedirs = ".git* templates static"
 python_files = "test_*.py"
